32-bit: download correct files, detect arch

+++ autoupdate.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-BINARY_LINK='https://zoom.us/client/latest/zoom_x86_64.tar.xz'
+BINARY_LINK_64='https://zoom.us/client/latest/zoom_x86_64.tar.xz'
 ARCH="$(arch)"
 
 # define error function
@@ -52,14 +52,20 @@ function setup-zoom() {
     rm -f $HOME/.local/share/applications/zoom.desktop
     rm -f $HOME/Desktop/zoom.desktop
     cd $HOME || error "Failed to navigate to user home directory."
-    wget $BINARY_LINK -O zoom.tar.xz || error "Failed to download Zoom archive."
-    tar -xvf zoom.tar.xz || error "Failed to extract Zoom archive."
-    rm zoom.tar.xz || error "Failed to remove zoom archive, as it isn't needed anymore."
-    wget https://github.com/ryanfortner/ZoomClient-ARM/raw/master/zoom_x64_libs.zip || error "Failed to download zoom x64 libraries!"
-    unzip zoom_x64_libs.zip || error "Failed to extract zoom libraries."
-    mv zoom_x64_libs/* zoom/ || error "Failed to move zoom x64 libraries to zoom folder."
-    rm -r $HOME/zoom_x64_libs || error "Failed to remove library folder."
-    cd $HOME/zoom && wget https://github.com/ryanfortner/ZoomClient-ARM/raw/master/icon.png || error "Failed to download icon."
+    if [ ! -z "$(file "$(readlink -f "/sbin/init")" | grep 64)" ];then
+      echo "downloading zoom x86_64..."
+      wget $BINARY_LINK_64 -O zoom.tar.xz || error "Failed to download zoom for arm64!"
+      wget https://github.com/ryanfortner/rpi-zoom/raw/master/zoom_x64_libs.zip || error "Failed to download zoom x64 libraries!"
+      unzip zoom_x64_libs.zip || error "Failed to extract zoom libraries."
+      mv zoom_x64_libs/* zoom/ || error "Failed to move zoom x64 libraries to zoom folder."
+      rm -r $HOME/zoom_x64_libs || error "Failed to remove library folder."
+    elif [ ! -z "$(file "$(readlink -f "/sbin/init")" | grep 32)" ];then
+      echo "downloading zoom for 32-bit..."
+      wget 'https://zoom.us/client/5.4.53391.1108/zoom_i686.tar.xz' -O zoom.tar.xz || wget 'https://zoom.com/client/5.4.53391.1108/zoom_i686.tar.xz' -O zoom.tar.xz || wget 'https://d11yldzmag5yn.cloudfront.net/prod/5.4.53391.1108/zoom_i686.tar.xz' -O zoom.tar.xz|| error 'Failed to download Zoom i686!'
+    else
+        error "Failed to detect architecture. Exiting..."
+    fi
+    cd $HOME/zoom && wget https://github.com/ryanfortner/rpi-zoom/raw/master/icon.png || error "Failed to download icon."
     echo "[Desktop Entry]
 Name=Zoom
 Exec=$HOME/zoom/zoom
